Follow the safer DataMapper style where only gems that Merb owns are wiped.
The way it was previously these sake tasks would wipe merb gems that it did
not own, such as merb_datamapper which now belongs to the DataMapper project.

I would also remove gems that had merb in the name in a pattern that matched
the regex.

Also modified it so that it will uninstall all Merb core, more, plugins gems
and re-install all of these by default with the merb:install command.  If you
want to install only a subset you should run one of merb:install:SOMETHING